+++ mmdet3d/core/__init__.py
@@ -8,3 +8,4 @@ from .post_processing import *  # noqa: F401, F403
 from .utils import *  # noqa: F401, F403
 from .visualizer import *  # noqa: F401, F403
 from .voxel import *  # noqa: F401, F403
+from .two_stage_runner import *