+++ mmdet3d/core/bbox/assigners/__init__.py
@@ -1,4 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from mmdet.core.bbox import AssignResult, BaseAssigner, MaxIoUAssigner
-
-__all__ = ['BaseAssigner', 'MaxIoUAssigner', 'AssignResult']
+from .hungarian_assigner_3d import HungarianAssigner3D
+__all__ = ['BaseAssigner', 'MaxIoUAssigner', 'AssignResult',
+           'HungarianAssigner3D']