+++ mmdet3d/datasets/pipelines/loading.py
@@ -1,6 +1,7 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 import os
 
+import cv2
 import mmcv
 import numpy as np
 import torch
@@ -11,6 +12,7 @@ from mmdet3d.core.points import BasePoints, get_points_type
 from mmdet.datasets.pipelines import LoadAnnotations, LoadImageFromFile
 from ...core.bbox import LiDARInstance3DBoxes
 from ..builder import PIPELINES
+
 
 @PIPELINES.register_module()
 class LoadOccGTFromFile(object):
@@ -26,8 +28,6 @@ class LoadOccGTFromFile(object):
         results['voxel_semantics'] = semantics
         results['mask_lidar'] = mask_lidar
         results['mask_camera'] = mask_camera
-
-
         return results
 
 
@@ -802,6 +802,49 @@ class PointToMultiViewDepth(object):
         return results
 
 
+@PIPELINES.register_module()
+class PointToMultiViewDepthFusion(PointToMultiViewDepth):
+    def __call__(self, results):
+        points_camego_aug = results['points'].tensor[:, :3]
+        # print(points_lidar.shape)
+        imgs, rots, trans, intrins = results['img_inputs'][:4]
+        post_rots, post_trans, bda = results['img_inputs'][4:]
+        points_camego = points_camego_aug - bda[:3, 3].view(1,3)
+        points_camego = points_camego.matmul(torch.inverse(bda[:3,:3]).T)
+
+        depth_map_list = []
+        for cid in range(len(results['cam_names'])):
+            cam_name = results['cam_names'][cid]
+
+            cam2camego = np.eye(4, dtype=np.float32)
+            cam2camego[:3, :3] = Quaternion(
+                results['curr']['cams'][cam_name]
+                ['sensor2ego_rotation']).rotation_matrix
+            cam2camego[:3, 3] = results['curr']['cams'][cam_name][
+                'sensor2ego_translation']
+            cam2camego = torch.from_numpy(cam2camego)
+
+            cam2img = np.eye(4, dtype=np.float32)
+            cam2img = torch.from_numpy(cam2img)
+            cam2img[:3, :3] = intrins[cid]
+
+            camego2img = cam2img.matmul(torch.inverse(cam2camego))
+
+            points_img = points_camego.matmul(
+                camego2img[:3, :3].T) + camego2img[:3, 3].unsqueeze(0)
+            points_img = torch.cat(
+                [points_img[:, :2] / points_img[:, 2:3], points_img[:, 2:3]],
+                1)
+            points_img = points_img.matmul(
+                post_rots[cid].T) + post_trans[cid:cid + 1, :]
+            depth_map = self.points2depthmap(points_img, imgs.shape[2],
+                                             imgs.shape[3])
+            depth_map_list.append(depth_map)
+        depth_map = torch.stack(depth_map_list)
+        results['gt_depth'] = depth_map
+        return results
+
+
 def mmlabNormalize(img):
     from mmcv.image.photometric import imnormalize
     mean = np.array([123.675, 116.28, 103.53], dtype=np.float32)
@@ -829,11 +872,13 @@ class PrepareImageInputs(object):
         data_config,
         is_train=False,
         sequential=False,
+        opencv_pp=False,
     ):
         self.is_train = is_train
         self.data_config = data_config
         self.normalize_img = mmlabNormalize
         self.sequential = sequential
+        self.opencv_pp = opencv_pp
 
     def get_rot(self, h):
         return torch.Tensor([
@@ -844,7 +889,8 @@ class PrepareImageInputs(object):
     def img_transform(self, img, post_rot, post_tran, resize, resize_dims,
                       crop, flip, rotate):
         # adjust image
-        img = self.img_transform_core(img, resize_dims, crop, flip, rotate)
+        if not self.opencv_pp:
+            img = self.img_transform_core(img, resize_dims, crop, flip, rotate)
 
         # post-homography transformation
         post_rot *= resize
@@ -859,8 +905,20 @@ class PrepareImageInputs(object):
         b = A.matmul(-b) + b
         post_rot = A.matmul(post_rot)
         post_tran = A.matmul(post_tran) + b
-
+        if self.opencv_pp:
+            img = self.img_transform_core_opencv(img, post_rot, post_tran, crop)
         return img, post_rot, post_tran
+
+    def img_transform_core_opencv(self, img, post_rot, post_tran,
+                                  crop):
+        img = np.array(img).astype(np.float32)
+        img = cv2.warpAffine(img,
+                             np.concatenate([post_rot,
+                                            post_tran.reshape(2,1)],
+                                            axis=1),
+                             (crop[2]-crop[0], crop[3]-crop[1]),
+                             flags=cv2.INTER_LINEAR)
+        return img
 
     def img_transform_core(self, img, resize_dims, crop, flip, rotate):
         # adjust image
@@ -889,12 +947,21 @@ class PrepareImageInputs(object):
             resize += np.random.uniform(*self.data_config['resize'])
             resize_dims = (int(W * resize), int(H * resize))
             newW, newH = resize_dims
-            crop_h = int((1 - np.random.uniform(*self.data_config['crop_h'])) *
+            random_crop_height = \
+                self.data_config.get('random_crop_height', False)
+            if random_crop_height:
+                crop_h = int(np.random.uniform(max(0.3*newH, newH-fH),
+                                               newH-fH))
+            else:
+                crop_h = \
+                    int((1 - np.random.uniform(*self.data_config['crop_h'])) *
                          newH) - fH
             crop_w = int(np.random.uniform(0, max(0, newW - fW)))
             crop = (crop_w, crop_h, crop_w + fW, crop_h + fH)
             flip = self.data_config['flip'] and np.random.choice([0, 1])
             rotate = np.random.uniform(*self.data_config['rot'])
+            if self.data_config.get('vflip', False) and np.random.choice([0, 1]):
+                rotate += 180
         else:
             resize = float(fW) / float(W)
             if scale is not None:
@@ -932,6 +999,64 @@ class PrepareImageInputs(object):
         ego2global[:3, :3] = ego2global_rot
         ego2global[:3, -1] = ego2global_tran
         return sensor2ego, ego2global
+
+    def photo_metric_distortion(self, img, pmd):
+        """Call function to perform photometric distortion on images.
+        Args:
+            results (dict): Result dict from loading pipeline.
+        Returns:
+            dict: Result dict with images distorted.
+        """
+        if np.random.rand()>pmd.get('rate', 1.0):
+            return img
+
+        img = np.array(img).astype(np.float32)
+        assert img.dtype == np.float32, \
+            'PhotoMetricDistortion needs the input image of dtype np.float32,' \
+            ' please set "to_float32=True" in "LoadImageFromFile" pipeline'
+        # random brightness
+        if np.random.randint(2):
+            delta = np.random.uniform(-pmd['brightness_delta'],
+                                   pmd['brightness_delta'])
+            img += delta
+
+        # mode == 0 --> do random contrast first
+        # mode == 1 --> do random contrast last
+        mode = np.random.randint(2)
+        if mode == 1:
+            if np.random.randint(2):
+                alpha = np.random.uniform(pmd['contrast_lower'],
+                                       pmd['contrast_upper'])
+                img *= alpha
+
+        # convert color from BGR to HSV
+        img = mmcv.bgr2hsv(img)
+
+        # random saturation
+        if np.random.randint(2):
+            img[..., 1] *= np.random.uniform(pmd['saturation_lower'],
+                                          pmd['saturation_upper'])
+
+        # random hue
+        if np.random.randint(2):
+            img[..., 0] += np.random.uniform(-pmd['hue_delta'], pmd['hue_delta'])
+            img[..., 0][img[..., 0] > 360] -= 360
+            img[..., 0][img[..., 0] < 0] += 360
+
+        # convert color from HSV to BGR
+        img = mmcv.hsv2bgr(img)
+
+        # random contrast
+        if mode == 0:
+            if np.random.randint(2):
+                alpha = np.random.uniform(pmd['contrast_lower'],
+                                       pmd['contrast_upper'])
+                img *= alpha
+
+        # randomly swap channels
+        if np.random.randint(2):
+            img = img[..., np.random.permutation(3)]
+        return Image.fromarray(img.astype(np.uint8))
 
     def get_inputs(self, results, flip=None, scale=None):
         imgs = []
@@ -973,6 +1098,9 @@ class PrepareImageInputs(object):
             post_tran[:2] = post_tran2
             post_rot[:2, :2] = post_rot2
 
+            if self.is_train and self.data_config.get('pmd', None) is not None:
+                img = self.photo_metric_distortion(img, self.data_config['pmd'])
+
             canvas.append(np.array(img))
             imgs.append(self.normalize_img(img))
 
@@ -981,12 +1109,20 @@ class PrepareImageInputs(object):
                 for adj_info in results['adjacent']:
                     filename_adj = adj_info['cams'][cam_name]['data_path']
                     img_adjacent = Image.open(filename_adj)
-                    img_adjacent = self.img_transform_core(
-                        img_adjacent,
-                        resize_dims=resize_dims,
-                        crop=crop,
-                        flip=flip,
-                        rotate=rotate)
+                    if self.opencv_pp:
+                        img_adjacent = \
+                            self.img_transform_core_opencv(
+                                img_adjacent,
+                                post_rot[:2, :2],
+                                post_tran[:2],
+                                crop)
+                    else:
+                        img_adjacent = self.img_transform_core(
+                            img_adjacent,
+                            resize_dims=resize_dims,
+                            crop=crop,
+                            flip=flip,
+                            rotate=rotate)
                     imgs.append(self.normalize_img(img_adjacent))
             intrins.append(intrin)
             sensor2egos.append(sensor2ego)
@@ -1023,7 +1159,22 @@ class PrepareImageInputs(object):
 
 
 @PIPELINES.register_module()
-class LoadAnnotationsBEVDepth(object):
+class LoadAnnotations(object):
+
+    def __call__(self, results):
+        gt_boxes, gt_labels = results['ann_infos']
+        gt_boxes, gt_labels = torch.Tensor(gt_boxes), torch.tensor(gt_labels)
+        if len(gt_boxes) == 0:
+            gt_boxes = torch.zeros(0, 9)
+        results['gt_bboxes_3d'] = \
+            LiDARInstance3DBoxes(gt_boxes, box_dim=gt_boxes.shape[-1],
+                                 origin=(0.5, 0.5, 0.5))
+        results['gt_labels_3d'] = gt_labels
+        return results
+
+
+@PIPELINES.register_module()
+class BEVAug(object):
 
     def __init__(self, bda_aug_conf, classes, is_train=True):
         self.bda_aug_conf = bda_aug_conf
@@ -1037,15 +1188,18 @@ class LoadAnnotationsBEVDepth(object):
             scale_bda = np.random.uniform(*self.bda_aug_conf['scale_lim'])
             flip_dx = np.random.uniform() < self.bda_aug_conf['flip_dx_ratio']
             flip_dy = np.random.uniform() < self.bda_aug_conf['flip_dy_ratio']
+            translation_std = self.bda_aug_conf.get('tran_lim', [0.0, 0.0, 0.0])
+            tran_bda = np.random.normal(scale=translation_std, size=3).T
         else:
             rotate_bda = 0
             scale_bda = 1.0
             flip_dx = False
             flip_dy = False
-        return rotate_bda, scale_bda, flip_dx, flip_dy
+            tran_bda = np.zeros((1, 3), dtype=np.float32)
+        return rotate_bda, scale_bda, flip_dx, flip_dy, tran_bda
 
     def bev_transform(self, gt_boxes, rotate_angle, scale_ratio, flip_dx,
-                      flip_dy):
+                      flip_dy, tran_bda):
         rotate_angle = torch.tensor(rotate_angle / 180 * np.pi)
         rot_sin = torch.sin(rotate_angle)
         rot_cos = torch.cos(rotate_angle)
@@ -1074,28 +1228,36 @@ class LoadAnnotationsBEVDepth(object):
                 gt_boxes[:, 6] = -gt_boxes[:, 6]
             gt_boxes[:, 7:] = (
                 rot_mat[:2, :2] @ gt_boxes[:, 7:].unsqueeze(-1)).squeeze(-1)
+            gt_boxes[:, :3] = gt_boxes[:, :3] + tran_bda
         return gt_boxes, rot_mat
 
     def __call__(self, results):
-        gt_boxes, gt_labels = results['ann_infos']
-        gt_boxes, gt_labels = torch.Tensor(gt_boxes), torch.tensor(gt_labels)
-        rotate_bda, scale_bda, flip_dx, flip_dy = self.sample_bda_augmentation(
-        )
+        gt_boxes = results['gt_bboxes_3d'].tensor
+        gt_boxes[:,2] = gt_boxes[:,2] + 0.5*gt_boxes[:,5]
+        rotate_bda, scale_bda, flip_dx, flip_dy, tran_bda = \
+            self.sample_bda_augmentation()
         bda_mat = torch.zeros(4, 4)
         bda_mat[3, 3] = 1
         gt_boxes, bda_rot = self.bev_transform(gt_boxes, rotate_bda, scale_bda,
-                                               flip_dx, flip_dy)
+                                               flip_dx, flip_dy, tran_bda)
+        if 'points' in results:
+            points = results['points'].tensor
+            points_aug = (bda_rot @ points[:, :3].unsqueeze(-1)).squeeze(-1)
+            points[:,:3] = points_aug + tran_bda
+            points = results['points'].new_point(points)
+            results['points'] = points
         bda_mat[:3, :3] = bda_rot
+        bda_mat[:3, 3] = torch.from_numpy(tran_bda)
         if len(gt_boxes) == 0:
             gt_boxes = torch.zeros(0, 9)
         results['gt_bboxes_3d'] = \
             LiDARInstance3DBoxes(gt_boxes, box_dim=gt_boxes.shape[-1],
                                  origin=(0.5, 0.5, 0.5))
-        results['gt_labels_3d'] = gt_labels
-        imgs, rots, trans, intrins = results['img_inputs'][:4]
-        post_rots, post_trans = results['img_inputs'][4:]
-        results['img_inputs'] = (imgs, rots, trans, intrins, post_rots,
-                                 post_trans, bda_rot)
+        if 'img_inputs' in results:
+            imgs, rots, trans, intrins = results['img_inputs'][:4]
+            post_rots, post_trans = results['img_inputs'][4:]
+            results['img_inputs'] = (imgs, rots, trans, intrins, post_rots,
+                                     post_trans, bda_mat)
         if 'voxel_semantics' in results:
             if flip_dx:
                 results['voxel_semantics'] = results['voxel_semantics'][::-1,...].copy()