+++ mmdet3d/models/utils/__init__.py
@@ -4,8 +4,10 @@ from .edge_indices import get_edge_indices
 from .gen_keypoints import get_keypoints
 from .handle_objs import filter_outside_objs, handle_proj_objs
 from .mlp import MLP
+from .transformer import *
+from .grid_mask import GridMask
 
 __all__ = [
     'clip_sigmoid', 'MLP', 'get_edge_indices', 'filter_outside_objs',
-    'handle_proj_objs', 'get_keypoints'
+    'handle_proj_objs', 'get_keypoints', 'GridMask'
 ]