+++ docs/en/news.md
@@ -1,5 +1,9 @@
 ## News
-
+- **2023.05.07** Improve the occpancy baseline by enlarging the input size and using long-term temporal fusion.
+- **2023.04.28** Support task of [occupancy prediction](https://github.com/CVPR2023-3D-Occupancy-Prediction/CVPR2023-3D-Occupancy-Prediction) .
+- **2023.04.27** Equip BEVDet with stereo depth estimation.
+- **2023.04.10** Use single head for multi-class prediction.
+- **2023.01.12** Support TensorRT-INT8.
 * **2022.11.24** A new branch of bevdet codebase, dubbed dev2.0, is released. dev2.0 includes the following features:
 1. support **BEVPoolv2**, whose inference speed is up to **15.1 times** the previous fastest implementation of Lift-Splat-Shoot view transformer. It is also far less memory consumption.
  ![bevpoolv2](../../resources/bevpoolv2.png)